Check for column existence before removing it

+++ db/migrate/20191028202451_drop_search_column_from_metadata_profiles.rb
@@ -1,5 +1,7 @@
 class DropSearchColumnFromMetadataProfiles < ActiveRecord::Migration[5.2]
   def change
-    remove_column :metadata_profiles, :search
+    if column_exists? :metadata_profiles, :search
+      remove_column :metadata_profiles, :search
+    end
   end
 end
